+++ src/Agent/NewRelic/Agent/Core/Transactions/Transaction.cs
@@ -559,7 +559,7 @@ namespace NewRelic.Agent.Core.Transactions
                 {
                     return headers;
                 }
-                return headers.Concat(new[] { new KeyValuePair<string, string>(Constants.DistributedTracePayloadKey, payload.HttpSafe()) });
+                return headers.Concat(new[] { new KeyValuePair<string, string>(Constants.DistributedTracePayloadKeyAllLower, payload.HttpSafe()) });
             }
 
             // CAT
@@ -605,28 +605,6 @@ namespace NewRelic.Agent.Core.Transactions
         {
             var pathHash = Agent._pathHashMaker.CalculatePathHash(transactionMetricName.PrefixedName, TransactionMetadata.CrossApplicationReferrerPathHash);
             TransactionMetadata.SetCrossApplicationPathHash(pathHash);
-        }
-
-        public void AcceptDistributedTracePayload(string payload, TransportType transportType)
-        {
-            if (!_configuration.DistributedTracingEnabled)
-            {
-                return;
-            }
-            if (TransactionMetadata.HasOutgoingTraceHeaders)
-            {
-                Agent._agentHealthReporter.ReportSupportabilityDistributedTraceAcceptPayloadIgnoredCreateBeforeAccept();
-                return;
-            }
-            if (TracingState != null)
-            {
-                Agent._agentHealthReporter.ReportSupportabilityDistributedTraceAcceptPayloadIgnoredMultiple();
-                return;
-            }
-
-            var isUnknownTransportType = transportType < TransportType.Unknown || transportType > TransportType.Other;
-
-            TracingState = Agent._distributedTracePayloadHandler.AcceptDistributedTracePayload(payload, isUnknownTransportType ? TransportType.Unknown : transportType, StartTime);
         }
 
         public void AcceptDistributedTraceHeaders<T>(T carrier, Func<T, string, IEnumerable<string>> getter, TransportType transportType)